va-8813 footer path fix
