+++ components/footer/desktop.tsx
@@ -1,6 +1,6 @@
 import { recordEvent } from '@/lib/utils/recordEvent';
-import { FOOTER_EVENTS, FOOTER_COLUMNS, FOOTER_CONTENT } from './helpers';
-import { createLinkGroups } from '.';
+import { FOOTER_EVENTS, FOOTER_COLUMNS, FOOTER_CONTENT, createLinkGroups } from './helpers';
+
 
 export const DesktopLinks = ({ links, visible }) => {
     const desktop = createLinkGroups(links)
@@ -43,7 +43,7 @@ export const DesktopLinks = ({ links, visible }) => {
                 </h2>
                 {desktop[FOOTER_COLUMNS.CONTACT]}
             </div>
-        </div >
+        </div>
     );
 }
