+++ components/footer/helpers.tsx
@@ -27,3 +27,45 @@ export const FOOTER_CONTENT = {
 };
 
 
+export const renderColumns = (columnLinks) => {
+    const render = columnLinks.map(column => {
+        return (
+            <>
+                <ul className="va-footer-links">
+                    {column.links.map(link => {
+                        return (
+                            <li key={link.title}>
+                                <span className="va-footer-link-text"><a href={link.href}>{link.title}</a></span>
+                            </li>
+                        )
+                    })}
+                </ul>
+            </>
+        )
+    });
+    return render;
+}
+
+export const generateLinkItems = (allColumns) => {
+    const linkColumns = Object.keys(allColumns).map(key => {
+        return {
+            columns: key,
+            links: allColumns[key]
+        }
+    });
+    return renderColumns(linkColumns)
+}
+
+export const createLinkGroups = (links) => {
+    const allGroups = links.reduce((acc, link) => {
+        if (!acc[link.column]) {
+            acc[link.column] = [];
+        }
+        acc[link.column].push({
+            href: link.href,
+            title: link.title
+        });
+        return acc;
+    }, {});
+    return generateLinkItems(allGroups);
+}
