+++ components/layout/index.tsx
@@ -7,6 +7,7 @@ import { HEADER_FOOTER_PATH } from '@/lib/constants';
 
 export default function Layout({ children }) {
     const { data, error } = useSWR(HEADER_FOOTER_PATH, fetcher);
+
     if (error || !data) return null;
 
     const { footerData } = data;
